adding a confirmation service and confirmation modal

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
 import { MessageService } from '../_services/message.service';
+import { ConfirmService } from '../_services/confirm.service';
 
 @Component({
   selector: 'app-messages',
@@ -16,7 +17,10 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading = false;
 
-  constructor(private messageService: MessageService) {}
+  constructor(
+    private messageService: MessageService,
+    private confirmService: ConfirmService
+  ) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -36,14 +40,22 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe({
-      next: () => {
-        this.messages?.splice(
-          this.messages.findIndex((m) => (m.id = id)),
-          1
-        );
-      },
-    });
+    this.confirmService
+      .confirm('Confirm delete message', 'This cannot be undone')
+      .subscribe({
+        next: (result) => {
+          if (result) {
+            this.messageService.deleteMessage(id).subscribe({
+              next: () => {
+                this.messages?.splice(
+                  this.messages.findIndex((m) => (m.id = id)),
+                  1
+                );
+              },
+            });
+          }
+        },
+      });
   }
 
   pageChanged(event: any) {
